refactor(ui): add new CoinOverview (splited from EthOverview)
