+++ ui/components/app/wallet-overview/index.scss
@@ -39,7 +39,7 @@
   font-weight: 500;
 }
 
-.eth-overview {
+.coin-overview {
   &__balance {
     display: flex;
     flex-direction: column;
